Improved REST endpoint status codes, implemented edit user account endpoint with password changing

+++ src/main/java/net/fosforito/partido/api/UsersApi.java
@@ -6,6 +6,7 @@ import net.fosforito.partido.model.user.UserDTO;
 import net.fosforito.partido.model.user.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.inject.Inject;
 import javax.ws.rs.core.MediaType;
+import java.util.Optional;
 
 @RestController
 public class UsersApi {
@@ -32,7 +34,7 @@ public class UsersApi {
 
   @GetMapping(value = "/users", produces = MediaType.APPLICATION_JSON)
   @PreAuthorize("hasRole('ADMIN')")
-  public Page<User> usersGet(Pageable pageable) {
+  public Page<User> getUsers(Pageable pageable) {
     return userRepository.findAll(pageable);
   }
 
@@ -42,7 +44,7 @@ public class UsersApi {
   }
 
   @PostMapping(value = {"/users"}, produces = MediaType.APPLICATION_JSON, consumes = MediaType.APPLICATION_JSON)
-  public User usersPost(@RequestBody UserDTO userDTO) {
+  public User createUser(@RequestBody UserDTO userDTO) {
     User user = new User();
     user.setUsername(userDTO.getUsername());
     user.setEmail(userDTO.getEmail());
@@ -53,37 +55,51 @@ public class UsersApi {
 
   @DeleteMapping(value = "/users/{userId}")
   @PreAuthorize("@securityService.userIsSameUser(principal, #userId) OR hasRole('ADMIN')")
-  public ResponseEntity<?> usersUserIdDelete(@PathVariable Long userId) throws Exception {
+  public ResponseEntity<?> deleteUser(@PathVariable Long userId) {
     //TODO: what will be deleted? Or just anonymize data?
-    return userRepository.findById(userId)
-        .map(user -> {
-          userRepository.delete(user);
-          return ResponseEntity.ok().build();
-        }).orElseThrow(() -> new Exception("User not found with id " + userId));
+    Optional<User> userOptional = userRepository.findById(userId);
+    if (userOptional.isPresent()) {
+      userOptional.map(user -> {
+        userRepository.delete(user);
+        return ResponseEntity.ok().build();
+      });
+    }
+    return ResponseEntity.notFound().build();
   }
 
   /**
+   * ADMIN only
+   *
    * To see user details from a group member, the users shipped
    * with the getGroup-Request should be used.
    *
+   * To see own user's details, the /currentuser endpoint should be used
+   *
    * @param userId ID of the user to be fetched
-   * @return User object
+   * @return User object or null if not existing
    */
   @GetMapping(value = "/users/{userId}", produces = MediaType.APPLICATION_JSON)
-  @PreAuthorize("@securityService.userIsSameUser(principal, #userId) OR hasRole('ADMIN')")
-  public User usersUserIdGet(@PathVariable Long userId) {
-    return userRepository.findById(userId).get();
+  @PreAuthorize("hasRole('ADMIN')")
+  public ResponseEntity<User> getUser(@PathVariable Long userId) {
+    Optional<User> userOptional = userRepository.findById(userId);
+    return userOptional.map(user -> new ResponseEntity<>(user, HttpStatus.OK))
+            .orElseGet(() -> new ResponseEntity<>(null, HttpStatus.NOT_FOUND));
   }
 
   @PutMapping(value = "/users/{userId}", produces = MediaType.APPLICATION_JSON, consumes = MediaType.APPLICATION_JSON)
-  @PreAuthorize("@securityService.userIsSameUser(principal, #userId) OR hasRole('ADMIN')")
-  public User usersUserIdPut(@PathVariable Long userId, @RequestBody UserDTO userDTO) throws Exception {
-    return userRepository.findById(userId)
-        .map(user -> {
-          user.setUsername(userDTO.getUsername());
-          user.setEmail(userDTO.getEmail());
-          user.setPassword(userDTO.getPassword());
-          return userRepository.save(user);
-        }).orElseThrow(() -> new Exception("User not found with id " + userId));
+  @PreAuthorize("@securityService.userIsSameUser(principal, #userId)")
+  public ResponseEntity<User> updateUser(@PathVariable Long userId, @RequestBody UserDTO userDTO) {
+    Optional<User> userOptional = userRepository.findById(userId);
+    if (passwordEncoder.matches(userDTO.getPassword(), userOptional.get().getPassword())) {
+      return new ResponseEntity<>(userOptional.map(user -> {
+        user.setUsername(userDTO.getUsername());
+        user.setEmail(userDTO.getEmail());
+        if (userDTO.getNewPassword() != null) {
+          user.setPassword(passwordEncoder.encode(userDTO.getNewPassword()));
+        }
+        return userRepository.save(user);
+      }).get(), HttpStatus.OK);
+    }
+    return new ResponseEntity<>(userOptional.get(), HttpStatus.NOT_MODIFIED);
   }
 }
